removed unused splunk metrics reporting service, client and models
