+++ pom.xml
@@ -26,14 +26,6 @@
         <apache.poi.version>3.17</apache.poi.version>
     </properties>
 
-    <repositories>
-        <repository>
-            <id>splunk-artifactory</id>
-            <name>Splunk Releases</name>
-            <url>https://splunk.jfrog.io/splunk/ext-releases-local</url>
-        </repository>
-    </repositories>
-
     <dependencies>
 
         <!-- commons -->
@@ -70,12 +62,6 @@
             <groupId>org.apache.xmlgraphics</groupId>
             <artifactId>xmlgraphics-commons</artifactId>
             <version>2.4</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.splunk</groupId>
-            <artifactId>splunk</artifactId>
-            <version>1.5.0.0</version>
         </dependency>
 
         <!-- Fasterxml Jackson dependencies - overrides vulnerable transitive dependencies
